리뷰 삭제 기능 추가

+++ src/main/java/com/zerobase/project6/store/domain/repository/StoreRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
     Optional<Store> findByStoreToken(String storeToken);
 }
